Create Separate component for react table and react pagination.
