+++ db.json
@@ -18,7 +18,7 @@
       "email": "yashbhatt078@gmail.com",
       "password": "22222222",
       "confirmPassword": "22222222",
-      "isVerified": true,
+      "isVerified": false,
       "isDeleted": true,
       "createdAt": "July 17th 2023",
       "updatedAt": "July 17th 2023",
@@ -31,7 +31,7 @@
       "email": "london@gmail.com",
       "password": "55555555",
       "confirmPassword": "55555555",
-      "isVerified": true,
+      "isVerified": false,
       "isDeleted": true,
       "createdAt": "July 17th 2023",
       "updatedAt": "July 17th 2023",
